new:usr: Working support to run tasks and report statistics
Implements a scheduled thread to report statistics from all tasks.
A thread is assigned a task. This is probably not up to protocol.
The weight determines the number of threads assigned to a task.

The java app accepts a class that implements a TaskList. The class
should be in classpath.